OM-556: Large values in app state slow down transactions
This patch introduces modifications to the parser's `path-meta` such
that it uses the query to walk the resulting data structure, leaving
pieces of data not present in the query as they were.

Since some utilities needed by the new `path-meta` were previously
present in the `om.next` namespace, they were moved to `om.util`.