OM-799: Make a new public function, `dom/create-element`
This is useful to create tags not present in `React.DOM` and therefore
not supported as vars in the `om.dom` namespace. Akin to calling
`js/React.createElement`, but more importantly to provide a way to
create these elements when doing server-side rendering, instead of
dropping to low-level functions like `om.dom/element`.