Remove apostrophe test, group encode/decode tests by subject
The apostrophe test that was previously included lost its value
due to an apostrophe being no special character when encoded as
base64 string. The base case + emoji case + naughty string case is
plenty to assert its correct function.

Furthermore, the tests have been grouped by subject so that the
encode/decode relation is more clearly expressed. In a single
reader conditional, the :clj is _encoding_ a string and putting it
in the DOM, and the :cljs is taking that same string (implicitly)
from the DOM and _decoding_ it.