Revert order merge-ident and merge-tree order again.
When I was developing I found the bug that's demonstrated by the new test case, the result with current code is that the new data from the child is not merged when its a placeholder. After some testing I realised that the original order now works fine on both previous failure case and this new one, due to the mark/sweep removal on the pre-merge step. So this can now work as it was in Fulcro 2.