file-upload: add response-type to outgoing request
This patch fixes a bug which let's the default response middleware not
extract the response originated by a file-upload originated request.

While introducing support for mutations returning binary data in commit
0ceb296cb34a380, a check was introduced to match up the response-type
of the request with that of the incoming response. Because file-upload
doesn't set response-type, the transit decoding was bypassed.