:replace: clarify that for vectors we need index (#456)
It was not clear to me what a "vector element" was. I thought perhaps it is an ident - since we have a vector of idents - and was confused at what sense that had. After reading the code I saw it actually needs to be an index and suddenly it made sense. For future readers this change will hopefully make it clear straight away.