Check query valid EQL at compile-time
The query we see at compile time is unresolved and may contain raw function calls etc.
We try to simplify it to something that can be checked, substituting calls to get-query
with a made-up query. If the query contains some other function calls (or, actually,
any other symbols outside of get-query calls) then we do not check it.