Bad prop type log improvement
Make sure we log the name of the cljs data structure type when
it is that (or, as we do now, "function ..." otherwise).

The `type` call just returns a reference to the `.-constructor`
and `type->str` gets the type name from it (at least in dev
mode, likely munged in release). If we just print the `type` as
we do now, it - at least sometimes - gets printed as the body of
the constructor instead of its cljs name. (Notice that the name is
not stored in its `.-name` prop as one might expect.)