Make sure fulcro app prints safely (#551)
* Make sure fulcro app prints safely

Issue: Printing a fulcro-app could blow the stack because an atom could contain itself. And even if it didn't, the app is huge and you rarely really want to print it.

Fix: Take inspiration from Component and replace the map with a record, for which we implement safe print-method. As a bonus, this also simplifies detecting whether something is a Fulcro app.

* Impl. also toString so that `(str app)` gives the same result.