Improve error message when a map passed by mistake as a child to fragment
... or to ErrorBoundary (which uses comp/fragment).

Without this change, we assume the map is props, and end up getting a _warning_ like "Warning: Invalid prop `some-prop...` supplied to `React.Fragment`. React.Fragment can only have `key` and `children` props."

With this change, we get a proper _error_ about trying to render the wrong type, and in which component.